"moved keybindings back to editor and added handleInterceptBackspace to shared"
