+++ src/eval-frame/index.jsx
@@ -20,7 +20,7 @@ import messagePasserEval from "../shared/utils/redux-to-port-message-passer";
 import { iodide } from "./iodide-api/api";
 
 import "./port-to-editor";
-import { initializeDefaultKeybindings } from "../shared/keybindings";
+import { handleInterceptBackspace } from "../shared/intercept-keybindings";
 
 messagePasserEval.connectDispatch(store.dispatch);
 window.iodide = iodide;
@@ -44,4 +44,4 @@ render(
 );
 
 initializeUserVariables();
-initializeDefaultKeybindings();
+handleInterceptBackspace();
