+++ src/shared/intercept-keybindings.js
@@ -1,5 +1,4 @@
 import Mousetrap from "mousetrap";
-import tasks from "../editor/user-tasks/task-definitions";
 
 Mousetrap.prototype.stopCallback = () => false;
 
@@ -9,13 +8,7 @@ const preventBacknav = e => {
   warnUser = e.target === document.body;
 };
 
-export function initializeDefaultKeybindings() {
-  Object.keys(tasks).forEach(t => {
-    const task = tasks[t];
-    if (task.hasKeybinding()) {
-      Mousetrap.bind(task.keybindings, task.keybindingCallback);
-    }
-  });
+export function handleInterceptBackspace() {
   Mousetrap.bind(["delete", "backspace"], preventBacknav);
 }
